Disabling Predictable network device naming for ubuntu1604
This fixes an issue with using ubuntu1604 with vagrant on VMware. When cloning a vagrant machine, vagrant will change the pci address which is used for the network card. When this happens and predicatble naming is on, the vagrant box will loose network connectivity.

This change reverts the new 1604 behaviour to that older versions of ubuntu.

Credits to: https://github.com/maglub/soe-ubuntu for the necessary changes